# Improve performance of `VideoStreamFrame.on_frame`