--- a/src/instamatic/config/camera/simulate.yaml
+++ b/src/instamatic/config/camera/simulate.yaml
@@ -12,7 +12,7 @@ camera_rotation_vs_stage_xy: -2.24
 default_binsize: 1
 default_exposure: 0.1
 dimensions: [516, 516]
-dynamic_range: 11800
+dynamic_range: 256
 interface: simulate
 physical_pixelsize: 0.055
 possible_binsizes: [1]